Yt-dlp added (100+ sites)

+++ src/main/java/works/akus/mediathief/stealer/DownloadTask.java
@@ -6,7 +6,7 @@ public interface DownloadTask {
 
     void onFailed(Throwable throwable);
 
-    void onProgress(int i);
+    void onProgress(float i);
 
     void onComplete(File file);
